+++ app/views/time_entry_reports/report.html.erb
@@ -7,52 +7,52 @@
 <h2><%= l(:label_spent_time) %></h2>
 
 <% form_tag({:controller => 'time_entry_reports', :action => 'report', :project_id => @project, :issue_id => @issue}, :method => :get, :id => 'query_form') do %>
-  <% @criterias.each do |criteria| %>
-    <%= hidden_field_tag 'criterias[]', criteria, :id => nil %>
+  <% @report.criteria.each do |criterion| %>
+    <%= hidden_field_tag 'criteria[]', criterion, :id => nil %>
   <% end %>
   <%= render :partial => 'timelog/date_range' %>
 
   <p><label for='columns'><%= l(:label_details) %></label>: <%= select_tag 'columns', options_for_select([[l(:label_year), 'year'],
                                                                             [l(:label_month), 'month'],
                                                                             [l(:label_week), 'week'],
-                                                                            [l(:label_day_plural).titleize, 'day']], @columns),
+                                                                            [l(:label_day_plural).titleize, 'day']], @report.columns),
                                                         :onchange => "this.form.onsubmit();" %>
 
-  <label for='criterias'><%= l(:button_add) %></label>: <%= select_tag('criterias[]', options_for_select([[]] + (@available_criterias.keys - @criterias).collect{|k| [l_or_humanize(@available_criterias[k][:label]), k]}),
+  <label for='criterias'><%= l(:button_add) %></label>: <%= select_tag('criteria[]', options_for_select([[]] + (@report.available_criteria.keys - @report.criteria).collect{|k| [l_or_humanize(@report.available_criteria[k][:label]), k]}),
                                                           :onchange => "this.form.submit();",
                                                           :style => 'width: 200px',
                                                           :id => nil,
-                                                          :disabled => (@criterias.length >= 3), :id => "criterias") %>
-     <%= link_to l(:button_clear), {:project_id => @project, :issue_id => @issue, :period_type => params[:period_type], :period => params[:period], :from => @from, :to => @to, :columns => @columns}, :class => 'icon icon-reload' %></p>
+                                                          :disabled => (@report.criteria.length >= 3), :id => "criterias") %>
+     <%= link_to l(:button_clear), {:project_id => @project, :issue_id => @issue, :period_type => params[:period_type], :period => params[:period], :from => @from, :to => @to, :columns => @report.columns}, :class => 'icon icon-reload' %></p>
 <% end %>
 
-<% unless @criterias.empty? %>
+<% unless @report.criteria.empty? %>
 <div class="total-hours">
-<p><%= l(:label_total) %>: <%= html_hours(l_hours(@total_hours)) %></p>
+<p><%= l(:label_total) %>: <%= html_hours(l_hours(@report.total_hours)) %></p>
 </div>
 
-<% unless @hours.empty? %>
+<% unless @report.hours.empty? %>
 <table class="list" id="time-report">
 <thead>
 <tr>
-<% @criterias.each do |criteria| %>
-  <th><%= l_or_humanize(@available_criterias[criteria][:label]) %></th>
+<% @report.criteria.each do |criteria| %>
+  <th><%= l_or_humanize(@report.available_criteria[criteria][:label]) %></th>
 <% end %>
-<% columns_width = (40 / (@periods.length+1)).to_i %>
-<% @periods.each do |period| %>
+<% columns_width = (40 / (@report.periods.length+1)).to_i %>
+<% @report.periods.each do |period| %>
   <th class="period" width="<%= columns_width %>%"><%= period %></th>
 <% end %>
   <th class="total" width="<%= columns_width %>%"><%= l(:label_total) %></th>
 </tr>
 </thead>
 <tbody>
-<%= render :partial => 'report_criteria', :locals => {:criterias => @criterias, :hours => @hours, :level => 0} %>
+<%= render :partial => 'report_criteria', :locals => {:criterias => @report.criteria, :hours => @report.hours, :level => 0} %>
   <tr class="total">
   <td><%= l(:label_total) %></td>
-  <%= '<td></td>' * (@criterias.size - 1) %>
+  <%= '<td></td>' * (@report.criteria.size - 1) %>
   <% total = 0 -%>
-  <% @periods.each do |period| -%>
-    <% sum = sum_hours(select_hours(@hours, @columns, period.to_s)); total += sum -%>
+  <% @report.periods.each do |period| -%>
+    <% sum = sum_hours(select_hours(@report.hours, @report.columns, period.to_s)); total += sum -%>
     <td class="hours"><%= html_hours("%.2f" % sum) if sum > 0 %></td>
   <% end -%>
   <td class="hours"><%= html_hours("%.2f" % total) if total > 0 %></td>
