+++ test/functional/time_entry_reports_controller_test.rb
@@ -43,72 +43,71 @@ class TimeEntryReportsControllerTest < ActionController::TestCase
   end
 
   def test_report_all_projects_one_criteria
-    get :report, :columns => 'week', :from => "2007-04-01", :to => "2007-04-30", :criterias => ['project']
+    get :report, :columns => 'week', :from => "2007-04-01", :to => "2007-04-30", :criteria => ['project']
     assert_response :success
     assert_template 'report'
-    assert_not_nil assigns(:total_hours)
-    assert_equal "8.65", "%.2f" % assigns(:total_hours)
+    assert_not_nil assigns(:report)
+    assert_equal "8.65", "%.2f" % assigns(:report).total_hours
   end
 
   def test_report_all_time
-    get :report, :project_id => 1, :criterias => ['project', 'issue']
+    get :report, :project_id => 1, :criteria => ['project', 'issue']
     assert_response :success
     assert_template 'report'
-    assert_not_nil assigns(:total_hours)
-    assert_equal "162.90", "%.2f" % assigns(:total_hours)
+    assert_not_nil assigns(:report)
+    assert_equal "162.90", "%.2f" % assigns(:report).total_hours
   end
 
   def test_report_all_time_by_day
-    get :report, :project_id => 1, :criterias => ['project', 'issue'], :columns => 'day'
+    get :report, :project_id => 1, :criteria => ['project', 'issue'], :columns => 'day'
     assert_response :success
     assert_template 'report'
-    assert_not_nil assigns(:total_hours)
-    assert_equal "162.90", "%.2f" % assigns(:total_hours)
+    assert_not_nil assigns(:report)
+    assert_equal "162.90", "%.2f" % assigns(:report).total_hours
     assert_tag :tag => 'th', :content => '2007-03-12'
   end
 
   def test_report_one_criteria
-    get :report, :project_id => 1, :columns => 'week', :from => "2007-04-01", :to => "2007-04-30", :criterias => ['project']
+    get :report, :project_id => 1, :columns => 'week', :from => "2007-04-01", :to => "2007-04-30", :criteria => ['project']
     assert_response :success
     assert_template 'report'
-    assert_not_nil assigns(:total_hours)
-    assert_equal "8.65", "%.2f" % assigns(:total_hours)
+    assert_not_nil assigns(:report)
+    assert_equal "8.65", "%.2f" % assigns(:report).total_hours
   end
 
-  def test_report_two_criterias
-    get :report, :project_id => 1, :columns => 'month', :from => "2007-01-01", :to => "2007-12-31", :criterias => ["member", "activity"]
+  def test_report_two_criteria
+    get :report, :project_id => 1, :columns => 'month', :from => "2007-01-01", :to => "2007-12-31", :criteria => ["member", "activity"]
     assert_response :success
     assert_template 'report'
-    assert_not_nil assigns(:total_hours)
-    assert_equal "162.90", "%.2f" % assigns(:total_hours)
+    assert_not_nil assigns(:report)
+    assert_equal "162.90", "%.2f" % assigns(:report).total_hours
   end
 
   def test_report_one_day
-    get :report, :project_id => 1, :columns => 'day', :from => "2007-03-23", :to => "2007-03-23", :criterias => ["member", "activity"]
+    get :report, :project_id => 1, :columns => 'day', :from => "2007-03-23", :to => "2007-03-23", :criteria => ["member", "activity"]
     assert_response :success
     assert_template 'report'
-    assert_not_nil assigns(:total_hours)
-    assert_equal "4.25", "%.2f" % assigns(:total_hours)
+    assert_not_nil assigns(:report)
+    assert_equal "4.25", "%.2f" % assigns(:report).total_hours
   end
 
   def test_report_at_issue_level
-    get :report, :project_id => 1, :issue_id => 1, :columns => 'month', :from => "2007-01-01", :to => "2007-12-31", :criterias => ["member", "activity"]
+    get :report, :project_id => 1, :issue_id => 1, :columns => 'month', :from => "2007-01-01", :to => "2007-12-31", :criteria => ["member", "activity"]
     assert_response :success
     assert_template 'report'
-    assert_not_nil assigns(:total_hours)
-    assert_equal "154.25", "%.2f" % assigns(:total_hours)
+    assert_not_nil assigns(:report)
+    assert_equal "154.25", "%.2f" % assigns(:report).total_hours
     assert_tag :form,
       :attributes => {:action => "/projects/ecookbook/issues/1/time_entries/report", :id => 'query_form'}
   end
 
   def test_report_custom_field_criteria
-    get :report, :project_id => 1, :criterias => ['project', 'cf_1', 'cf_7']
+    get :report, :project_id => 1, :criteria => ['project', 'cf_1', 'cf_7']
     assert_response :success
     assert_template 'report'
-    assert_not_nil assigns(:total_hours)
-    assert_not_nil assigns(:criterias)
-    assert_equal 3, assigns(:criterias).size
-    assert_equal "162.90", "%.2f" % assigns(:total_hours)
+    assert_not_nil assigns(:report)
+    assert_equal 3, assigns(:report).criteria.size
+    assert_equal "162.90", "%.2f" % assigns(:report).total_hours
     # Custom field column
     assert_tag :tag => 'th', :content => 'Database'
     # Custom field row
@@ -121,15 +120,15 @@ class TimeEntryReportsControllerTest < ActionController::TestCase
   end
 
   def test_report_one_criteria_no_result
-    get :report, :project_id => 1, :columns => 'week', :from => "1998-04-01", :to => "1998-04-30", :criterias => ['project']
+    get :report, :project_id => 1, :columns => 'week', :from => "1998-04-01", :to => "1998-04-30", :criteria => ['project']
     assert_response :success
     assert_template 'report'
-    assert_not_nil assigns(:total_hours)
-    assert_equal "0.00", "%.2f" % assigns(:total_hours)
+    assert_not_nil assigns(:report)
+    assert_equal "0.00", "%.2f" % assigns(:report).total_hours
   end
 
   def test_report_status_criterion
-    get :report, :project_id => 1, :criterias => ['status']
+    get :report, :project_id => 1, :criteria => ['status']
     assert_response :success
     assert_template 'report'
     assert_tag :tag => 'th', :content => 'Status'
@@ -137,7 +136,8 @@ class TimeEntryReportsControllerTest < ActionController::TestCase
   end
 
   def test_report_all_projects_csv_export
-    get :report, :columns => 'month', :from => "2007-01-01", :to => "2007-06-30", :criterias => ["project", "member", "activity"], :format => "csv"
+    get :report, :columns => 'month', :from => "2007-01-01", :to => "2007-06-30",
+        :criteria => ["project", "member", "activity"], :format => "csv"
     assert_response :success
     assert_equal 'text/csv', @response.content_type
     lines = @response.body.chomp.split("\n")
@@ -148,7 +148,9 @@ class TimeEntryReportsControllerTest < ActionController::TestCase
   end
 
   def test_report_csv_export
-    get :report, :project_id => 1, :columns => 'month', :from => "2007-01-01", :to => "2007-06-30", :criterias => ["project", "member", "activity"], :format => "csv"
+    get :report, :project_id => 1, :columns => 'month',
+        :from => "2007-01-01", :to => "2007-06-30",
+        :criteria => ["project", "member", "activity"], :format => "csv"
     assert_response :success
     assert_equal 'text/csv', @response.content_type
     lines = @response.body.chomp.split("\n")
@@ -157,5 +159,4 @@ class TimeEntryReportsControllerTest < ActionController::TestCase
     # Total row
     assert_equal 'Total,"","","","",154.25,8.65,"","",162.90', lines.last
   end
-
 end
